feat: Refactor account page components
